chore(features): move tests to source
