--- features/steps/.test/assets/deployment/Chart.yaml
@@ -1,3 +1,0 @@
-apiVersion: v1
-name: dummies
-version: 0.0.0
