--- features/steps/.test/assets/directory/file1
@@ -1,2 +1,0 @@
-Foo
-Bar
