--- features/steps/.test/assets/directory/file2
@@ -1,2 +1,0 @@
-Baz
-Qux
