+++ features/steps/connectors.test.js
@@ -13,7 +13,8 @@ const mock = { gherkin, boot: fixtures.mock.boot }
 
 jest.mock('@cucumber/cucumber', () => mock.gherkin)
 jest.mock('@toa.io/boot', () => mock.boot)
-require('../connectors')
+
+require('./connectors')
 
 beforeEach(() => {
   jest.clearAllMocks()
@@ -249,9 +250,9 @@ describe('Then the reply is received:', () => {
 })
 
 describe('When an event {label} is emitted with the payload:', () => {
-  const step = gherkin.steps.Wh('an event {label} is emitted with the payload:')
+  gherkin.steps.Wh('an event {label} is emitted with the payload:')
 
   it('should be', async () => undefined)
 })
 
-const COLLECTION = resolve(__dirname, '../.workspace/components/collection')
+const COLLECTION = resolve(__dirname, '.workspace/components/collection')
