+++ features/steps/hooks.test.js
@@ -4,7 +4,7 @@ const { directory } = require('@toa.io/filesystem')
 const mock = require('@toa.io/mock')
 
 jest.mock('@cucumber/cucumber', () => mock.gherkin)
-require('../hooks')
+require('./hooks')
 
 const gherkin = mock.gherkin
